requirements.txt
streamlit
openai
python-dotenv
python-docx
markdown
xhtml2pdf
PyPDF2
psycopg2